refactor : ♻️ index에서 실행했던 파트들을 controller로 이관 index.js에서 다루는 함수들의 개수가 늘어나 controller파일을 따로 생성하여 해당 파일에서 유저의 활동에 따라 화면을 렌더링하는 함수들과 유저의 클릭과 입력 이벤트를 처리하는 함수들을 생성하고 이들을  추가함  + 동일한 render 역할을 하는 함수들을 renderVendingMachineCoinInfo에서 호출해 코드를 간단하게함 index.js / controller.js / controlProductAddManage.js / controlProductPurchase.js / controlVendingMachineManage.js / setDrinkStorage.js / renderVendingMachineCoinInfo.js
