+++ src/controllers/controlProductPurchase.js
@@ -1,0 +1,1 @@
+export default function controlProductPurchase() {}
