+++ src/views/renderVendingMachineCoinInfo.js
@@ -1,0 +1,7 @@
+import renderTotalCoinElement from './renderTotalCoinElement.js';
+import renderVendingMachineOwnCoinTable from './renderVendingMachineOwnCoinTable.js';
+
+export default function renderVendingMachineCoinInfo() {
+  renderTotalCoinElement();
+  renderVendingMachineOwnCoinTable();
+}
